services/ml-serving/serving/services/gemini_service.py
import google.generativeai as genai
from ..utils import get_env

GEMINI_API_KEY = get_env("GEMINI_API_KEY")
if GEMINI_API_KEY:
    genai.configure(api_key=GEMINI_API_KEY)

async def ask_gemini_model(prompt: str) -> str:
    model = genai.GenerativeModel("gemini-2.5-pro")
    response = model.generate_content(prompt)
    return response.text
