services/ml-serving/requirements.txt
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2
xgboost==2.0.0
optuna==3.3.0
joblib==1.3.2
fastapi==0.103.1
uvicorn[standard]==0.23.2
requests==2.31.0
openpyxl==3.1.2
httpx==0.24.1
python-dotenv==1.0.0
fastapi-cache2==0.2.1
pydantic==2.3.0
xlrd==2.0.1
lxml==4.9.3
python-multipart==0.0.6
google-generativeai
